+++ pytest_notebook/plugin.py
@@ -8,6 +8,7 @@ For more information on writing pytest plugins see:
 - https://docs.pytest.org/en/latest/example/simple.html
 - https://github.com/pytest-dev/cookiecutter-pytest-plugin
 - http://doc.pytest.org/en/latest/example/nonpython.html
+- https://docs.pytest.org/en/latest/_modules/_pytest/hookspec.html
 
 """
 from distutils.util import strtobool as _str2bool
@@ -42,6 +43,7 @@ class NotSet:
 
 
 def pytest_addoption(parser):
+    """Add pytest commandline and configuration file options."""
     group = parser.getgroup("nb_regression")
     group.addoption(
         "--nb-test-files",
@@ -197,6 +199,20 @@ def gather_config_options(pytestconfig):
         nbreg_kwargs["diff_replace"] = nb_diff_replace
 
     return nbreg_kwargs, other_args
+
+
+def pytest_report_header(config):
+    """Add header information for pytest execution."""
+
+    kwargs, other_args = gather_config_options(config)
+    header = []
+    if kwargs.get("exec_notebook", True) and kwargs.get("exec_cwd", None):
+        header.append(f"NB exec dir: {kwargs['exec_cwd']}")
+    if kwargs.get("post_processors", None):
+        header.append(f"NB post processors: {' '.join(kwargs['post_processors'])}")
+    if kwargs.get("force_regen", None):
+        header.append(f"NB force regen: {kwargs['force_regen']}")
+    return header
 
 
 @pytest.fixture(scope="function")