+++ tests/test_postprocessors/test_blacken_code.py
@@ -1,7 +1,8 @@
+"""Tests for blacken_code post-processor."""
 import textwrap
 
 from pytest_notebook.post_processors import blacken_code
-from pytest_notebook.notebook import create_notebook, prepare_cell
+from pytest_notebook.notebook import create_notebook, mapping_to_dict, prepare_cell
 
 
 def test_blacken_no_code():
@@ -11,8 +12,8 @@ def test_blacken_no_code():
     assert notebook == new_notebook
 
 
-def test_blacken_code():
-    """Test coalesce_streams if streams require merging."""
+def test_blacken_code(data_regression):
+    """Test blacken unformatted code."""
     notebook = create_notebook()
     notebook.cells.append(
         prepare_cell(
@@ -31,23 +32,6 @@ def test_blacken_code():
             }
         )
     )
-    expected = create_notebook()
-    expected.cells.append(
-        prepare_cell(
-            {
-                "cell_type": "code",
-                "execution_count": 1,
-                "metadata": {},
-                "outputs": [],
-                "source": textwrap.dedent(
-                    """\
-                    for i in range(5):
-                        x = i
-                        a = "123"  # comment
-                    """
-                ),
-            }
-        )
-    )
+
     new_notebook, _ = blacken_code(notebook, {})
-    assert new_notebook == expected
+    data_regression.check(mapping_to_dict(new_notebook))